+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Changed
 
 * Updated cross icon for file property with new color [#158](https://github.com/wp-papi/papi/pull/158).
+* Updated so `$id` param for `papi_get_slugs` is optional.
 
 ### Removed